Load cards data async for website
